requirements.txt
django>=5.0
djangorestframework>=3.14
gunicorn>=21.2
pydantic-settings>=2.0
python-json-logger>=2.0

